requirements.txt
certifi==2020.12.5
mkl-fft==1.3.0
mkl-random==1.2.0
mkl-service==2.3.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.1.48
Pillow @ file:///tmp/build/80754af9/pillow_1615057347459/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
